+++ src/components/Header/Layouts/MobileHeader/AccountOverview/AccountOverview.scss
@@ -1,0 +1,47 @@
+.overviewContainer {
+  text-align: center;
+  margin-bottom: 30px;
+
+  .identicon {
+    width: 90px;
+    height: 90px;
+
+    border-radius: 50%;
+    border: 1px solid white;
+  }
+
+  .networkBalanceWrapper {
+    display: flex;
+    margin-bottom: 10px;
+
+    & > div {
+      line-height: 1;
+      margin-top: -1px;
+    }
+
+    span {
+      font-size: 15px;
+      font-weight: bold;
+    }
+
+    .network {
+      margin-right: auto;
+    }
+  }
+
+  .addressWrapper {
+    margin-top: 10px;
+    .address {
+      padding: 0;
+      border: 0;
+      background: transparent;
+      color: #b8b7ad;
+
+      &:focus,
+      &:hover {
+        color: darken($color: #b8b7ad, $amount: 20);
+        outline: 0;
+      }
+    }
+  }
+}
